+++ cosmospy/__init__.py
@@ -1,1 +1,9 @@
+from cosmospy._transaction import Transaction  # noqa: F401
+from cosmospy._wallet import (  # noqa: F401
+    generate_wallet,
+    privkey_to_address,
+    privkey_to_pubkey,
+    pubkey_to_address,
+)
+
 __version__ = "3.0.2"  # DO NOT EDIT THIS LINE MANUALLY. LET bump2version UTILITY DO IT